refactor: common controller 패키지 이름 presentation으로 변경
